+++ src/Core/src/ScopeInterface.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Spiral\Core;
 
-use Throwable;
-
 /**
  * Provides ability to run code withing isolated IoC scope.
  */
@@ -20,11 +18,7 @@ interface ScopeInterface
      *    dump($container->get('actor'));
      * });
      *
-     * @param array    $bindings
-     * @param callable $scope
-     * @return mixed
-     *
-     * @throws Throwable
+     * @throws \Throwable
      */
-    public function runScope(array $bindings, callable $scope);
+    public function runScope(array $bindings, callable $scope): mixed;
 }